+++ src/sisl/io/siesta/_help.py
@@ -21,7 +21,6 @@ except ImportError:
 __all__ = ["_csr_from_siesta", "_csr_from_sc_off"]
 __all__ += ["_csr_to_siesta", "_csr_to_sc_off"]
 __all__ += ["_mat_spin_convert", "_fc_correct"]
-__all__ += ["_replace_with_species"]
 
 
 def _ensure_diagonal(csr):
@@ -159,20 +158,6 @@ def _geom2hsx(geometry):
             l.append([-1 for orb in atom])
             zeta.append([1 for orb in atom])
     return (label, Z, no), (n, l, zeta)
-
-
-def _replace_with_species(basis, ref_basis):
-    """Replace the `basis` with the atoms in `ref_basis`
-
-    This method will assume that the `basis` contains
-    the atomic numbers as their species (i.e. starting from 1..n-species).
-    It will make it simple to change according to basis.
-    """
-    with warnings.catch_warnings():
-        warnings.simplefilter("ignore")
-        for atom, _ in basis.iter(True):
-            basis.replace(atom, ref_basis[atom.Z - 1])
-        basis.reduce(inplace=True)
 
 
 def _fc_correct(fc, trans_inv=True, sum0=True, hermitian=True):